catch null pointer

+++ app/src/main/java/ngo/teog/swift/gui/main/MainActivity.java
@@ -211,11 +211,7 @@ public class MainActivity extends BaseActivity {
      */
     public void logout() {
         ExecutorService executor = Executors.newFixedThreadPool(1);
-
-        executor.execute(() -> {
-            database.clearAllTables();
-        });
-
+        executor.execute(() -> database.clearAllTables());
         executor.shutdown();
 
         //delete shared preferences
@@ -228,8 +224,12 @@ public class MainActivity extends BaseActivity {
         File imageDir = new File(getFilesDir(), Defaults.DEVICE_IMAGE_PATH);
 
         if(imageDir.exists()) {
-            for(File file : imageDir.listFiles()) {
-                file.delete();
+            File[] files = imageDir.listFiles();
+
+            if(files != null) {
+                for (File file : files) {
+                    file.delete();
+                }
             }
         }
